% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_degree}
\alias{get_degree}
\title{Computes the score statistic}
\usage{
get_degree(x, y, Y)
}
\arguments{
\item{x}{covariate matrix to test}

\item{y}{vector of main variable}

\item{Y}{matrix of the rest of the variables. The row number of Y should match the length of y.}
}
\value{
double: degree statistic
}
\description{
Computes the score statistic
}
\examples{
n = 30
k = 3
nullX = as.matrix(rnorm(30), ncol=1)
nullY = MASS::mvrnorm(n, rep(0,3), diag(3))
d = get_degree(as.matrix(nullX, ncol=1), nullY[,1], nullY[,2:3])
p = get_p_from_degree(nullY[,1], nullY[,2:3], d)
print(paste("The degree statistic is", d, "and the p-value is", p))
}
