% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_est_H}
\alias{get_est_H}
\title{Computes the correlation matrix of the combined test statistics r
Takes input of the estimated or true covariance matrix}
\usage{
get_est_H(Sigma)
}
\arguments{
\item{Sigma}{estimate or true covariance matrix of $K$ variables}
}
\value{
estimated H from the covariance matrix Sigma
}
\description{
Computes the correlation matrix of the combined test statistics r
Takes input of the estimated or true covariance matrix
}
