% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_H}
\alias{get_H}
\title{Estimate H}
\usage{
get_H(Sigma)
}
\arguments{
\item{Sigma}{True or estimated variance of all the variables of interest.
The first row / column must be the main variable of interest}
}
\value{
Esitmated covariance matrix of the part of score statistics
}
\description{
Estimate H
}
