% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_newfisher}
\alias{get_newfisher}
\title{Computes the inverse Fisher information of the log likelihood}
\usage{
get_newfisher(x, y1, y2)
}
\arguments{
\item{x}{covariate matrix to test}

\item{y1}{vector of first variable}

\item{y2}{vector of second variable}
}
\value{
Matrix of inverse fisher information of the bivariate normal likelihood
}
\description{
Computes the inverse Fisher information of the log likelihood
}
