% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_newgrad}
\alias{get_newgrad}
\title{Computes the score of the log likelihood}
\usage{
get_newgrad(x, y1, y2)
}
\arguments{
\item{x}{covariate matrix to test}

\item{y1}{vector of first variable}

\item{y2}{vector of second variable}
}
\value{
score vector of the bivariate normal likelihood
}
\description{
Computes the score of the log likelihood
}
