% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_p_from_degree}
\alias{get_p_from_degree}
\title{Estimate p-value from the degree statistic}
\usage{
get_p_from_degree(y, Y, d, numsim = 5000)
}
\arguments{
\item{y}{vector: main variable of interest}

\item{Y}{Matrix: other variables of interest to measure the correlation with y}

\item{d}{double: computed degree statistic}

\item{numsim}{integer: number of simulations to draw the null distribution}
}
\value{
p value of the degree statistic
}
\description{
Estimate p-value from the degree statistic
}
\examples{
n = 30
k = 3
nullX = as.matrix(rnorm(30), ncol=1)
nullY = MASS::mvrnorm(n, rep(0,3), diag(3))
d = get_degree(as.matrix(nullX, ncol=1), nullY[,1], nullY[,2:3])
p = get_p_from_degree(nullY[,1], nullY[,2:3], d)
print(paste("The degree statistic is", d, "and the p-value is", p))
}
