% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R, R/utilities_wv.R
\name{post_score}
\alias{post_score}
\title{Computes the small sample correction}
\usage{
post_score(score, coef)

post_score(score, coef)
}
\arguments{
\item{score}{score statistic computed from get_score}

\item{coef}{coefficients for the cubic function}
}
\value{
small-sample adjusted score statistic

small-sample adjusted score statistic
}
\description{
Computes the small sample correction

Computes the small sample correction
}
